[CLANG] Fix [-Wpointer-to-enum-cast] warnings when pointer is type cast to enum
* See https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/1061#note_1037864522 and ensuing
  discussion for background.

* The issue is that a pointer type is being type cast to an enum. clang issues a warning about
  this while gcc does not.

* The fix is to type cast the pointer first to a `uintptr_t` type and then type cast it to the `enum`.

* There are various such usages fixed in this commit but all follow the same principle.

+++ src/octo.h
@@ -790,7 +790,7 @@ typedef enum RegexType {
 		OCTO_CMALLOC_STRUCT(ret->v.drop_table, SqlDropTableStatement);                                  \
 		ret->v.drop_table->table_name = COLUMN_NAME;                                                    \
 		ret->v.drop_table->optional_keyword = DROP_BEHAVIOR;                                            \
-		ret->v.drop_table->drop_data_retention = (enum OptionalKeyword)DROP_DATA_RETENTION;             \
+		ret->v.drop_table->drop_data_retention = (enum OptionalKeyword)(uintptr_t)DROP_DATA_RETENTION;  \
 		ret->v.drop_table->if_exists_specified = IF_EXISTS_SPECIFIED;                                   \
 		STMT = ret;                                                                                     \
 	}
